requirements.txt
fastapi==0.103.2
starlette==0.27.0
pydantic==1.10.24
tensorflow==2.13.0
tensorflow-intel==2.13.0 ; platform_system == "Windows"
numpy==1.24.3
pillow
python-multipart
protobuf==3.20.*
typing_extensions>=4.5.0,<4.6.0
